third_party/blink/web_tests/platform/generic/css3/tab-size-complex-path-expected.txt

Test tab-size measurements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS pxTabLength is pxExpected
PASS emTabLength is emExpected
PASS intTabLength is intExpected
PASS pxTabLength is pxExpected
PASS emTabLength is emExpected
PASS intTabLength is intExpected
PASS successfullyParsed is true

TEST COMPLETE
